Modifications required to get pip distribution working, new release
